+++ core/src/builtin_widgets.rs
@@ -75,6 +75,10 @@ pub use smooth_layout::*;
 
 mod track_widget_id;
 pub use track_widget_id::*;
+mod text;
+pub use text::*;
+mod tooltips;
+pub use tooltips::*;
 
 use crate::prelude::*;
 
@@ -131,6 +135,7 @@ pub struct FatObj<T> {
   painting_style: Option<State<PaintingStyleWidget>>,
   text_style: Option<State<TextStyleWidget>>,
   keep_alive: Option<State<KeepAlive>>,
+  tooltips: Option<State<Tooltips>>,
   keep_alive_unsubscribe_handle: Option<Box<dyn Any>>,
 }
 
@@ -165,6 +170,7 @@ impl<T> FatObj<T> {
       text_style: self.text_style,
       visibility: self.visibility,
       opacity: self.opacity,
+      tooltips: self.tooltips,
       keep_alive: self.keep_alive,
       keep_alive_unsubscribe_handle: self.keep_alive_unsubscribe_handle,
     }
@@ -195,6 +201,7 @@ impl<T> FatObj<T> {
       && self.visibility.is_none()
       && self.opacity.is_none()
       && self.keep_alive.is_none()
+      && self.tooltips.is_none()
   }
 
   /// Return the host object of the FatObj.
@@ -413,6 +420,14 @@ impl<T> FatObj<T> {
   pub fn get_keep_alive_widget(&mut self) -> &State<KeepAlive> {
     self
       .keep_alive
+      .get_or_insert_with(|| State::value(<_>::default()))
+  }
+
+  /// Returns the `State<Tooltips>` widget from the FatObj. If it doesn't
+  /// exist, a new one is created.
+  pub fn get_tooltips_widget(&mut self) -> &State<Tooltips> {
+    self
+      .tooltips
       .get_or_insert_with(|| State::value(<_>::default()))
   }
 }
@@ -857,6 +872,11 @@ impl<T> FatObj<T> {
     self.declare_builtin_init(v, Self::get_opacity_widget, |m, v| m.opacity = v)
   }
 
+  /// Initializes the tooltips of the widget.
+  pub fn tooltips<const M: u8>(self, v: impl DeclareInto<CowArc<str>, M>) -> Self {
+    self.declare_builtin_init(v, Self::get_tooltips_widget, |m, v| m.tooltips = v)
+  }
+
   /// Initializes the `keep_alive` value of the `KeepAlive` widget.
   pub fn keep_alive<const M: u8>(mut self, v: impl DeclareInto<bool, M>) -> Self {
     let (v, o) = v.declare_into().unzip();
@@ -944,6 +964,7 @@ impl<'a> FatObj<Widget<'a>> {
           class,
           cursor,
           constrained_box,
+          tooltips,
           margin,
           transform,
           opacity,